refactor: add short map names and update map ID type
Adds a `shortName` property to the `MapCarouselModel` to display abbreviated map names in the sanity tools carousel.

This commit also refactors the `mapId` from an `Int` to a `String` across various map-related models, including `MapModel`, `MapDesBlueprint`, and `MapListModel`, for better consistency. Additionally, a `getMapByIndex` method was added to `MapListModel` and logic in `NewsletterMessageModel` was simplified by removing premature XML tag stripping.